Added sorting, permission checks, edit functionality

+++ db/queries.ts
@@ -1,11 +1,13 @@
 "use server";
 
-import { eq } from "drizzle-orm";
+import { eq, inArray } from "drizzle-orm";
 import { db } from ".";
 import { SelectTask, tasks } from "./schema";
 import { Task, TaskSchema } from "@/types/task";
 import { z, ZodError } from "zod";
 import { isRateLimitExceeded } from "@/lib/rate-limiter";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
 
 const queryWrapper = async <T>(queryFunction: () => Promise<T>): Promise<T> => {
   try {
@@ -22,13 +24,30 @@ const queryWrapper = async <T>(queryFunction: () => Promise<T>): Promise<T> => {
 
 // Create
 export async function insertTask(data: Task): Promise<void> {
+  const session = await getServerSession(authOptions);
+
+  if (session?.user.id !== data.userId) {
+    throw new Error("Cannot insert a task for another user");
+  }
+
   await queryWrapper(async () => {
     await db.insert(tasks).values(TaskSchema.parse(data));
   });
 }
 
+export async function upsertTask(data: Task): Promise<void> {
+  if (!data.id) insertTask(data);
+  else updateTask(data.id, data);
+}
+
 // Read
 export async function getTasks(): Promise<SelectTask[]> {
+  const session = await getServerSession(authOptions);
+
+  if (session?.user.email !== "apaluchdev@gmail.com") {
+    throw new Error("Cannot read all tasks");
+  }
+
   return await queryWrapper(async () => {
     return await db.select().from(tasks);
   });
@@ -36,6 +55,12 @@ export async function getTasks(): Promise<SelectTask[]> {
 
 // Read
 export async function getTasksByUserId(userId: Task["userId"]): Promise<Task[]> {
+  const session = await getServerSession(authOptions);
+
+  if (session?.user.id !== userId) {
+    throw new Error("Cannot read tasks for another user");
+  }
+
   return await queryWrapper(async () => {
     const dbTasks: SelectTask[] = await db.select().from(tasks).where(eq(tasks.userId, userId));
     return dbTasks.map((t) => {
@@ -46,7 +71,13 @@ export async function getTasksByUserId(userId: Task["userId"]): Promise<Task[]> 
 
 // Update
 export async function updateTask(id: Task["id"], data: Partial<Omit<Task, "id">>) {
+  const session = await getServerSession(authOptions);
+
   if (!id) return;
+  if (session?.user.id !== data.userId) {
+    throw new Error("Cannot update a task for another user");
+  }
+
   data.updatedAt = new Date();
 
   return await queryWrapper(async () => {
@@ -58,7 +89,46 @@ export async function updateTask(id: Task["id"], data: Partial<Omit<Task, "id">>
 export async function deleteTask(id: Task["id"]) {
   if (!id) return;
 
+  const taskToDelete = (await db.select().from(tasks).where(eq(tasks.id, id)))[0];
+
+  if (!taskToDelete) return;
+
+  if (taskToDelete.userId !== (await getServerSession(authOptions))?.user.id) {
+    throw new Error("Cannot delete a task for another user");
+  }
+
   return await queryWrapper(async () => {
     await db.delete(tasks).where(eq(tasks.id, id));
+  });
+}
+
+// Delete multiple tasks
+export async function deleteTasks(ids: Task["id"][]) {
+  const session = await getServerSession(authOptions);
+  if (!ids || ids.length === 0) return;
+
+  const tasksToDelete = await db
+    .select()
+    .from(tasks)
+    .where(
+      inArray(
+        tasks.id,
+        ids.filter((id) => id !== undefined)
+      )
+    );
+
+  tasksToDelete.forEach((task) => {
+    if (task.userId !== session?.user.id) {
+      throw new Error("Cannot delete a task for another user");
+    }
+  });
+
+  return await queryWrapper(async () => {
+    await db.delete(tasks).where(
+      inArray(
+        tasks.id,
+        ids.filter((id) => id !== undefined)
+      )
+    );
   });
 }
